+++ package/src/extract/json.rs
@@ -1,12 +1,10 @@
-use axum::{
-    extract::Json as _Json,
-    http::{StatusCode, header},
-};
+use axum::extract::Json as _Json;
 use axum_core::{
     extract::{FromRequest, OptionalFromRequest, Request},
     response::IntoResponse,
 };
 use bytes::{BufMut, BytesMut, buf::Writer};
+use http::{StatusCode, header};
 use serde::{Serialize, de::DeserializeOwned};
 
 use crate::response::{