+++ package/src/extract/multipart/mod.rs
@@ -1,4 +1,7 @@
 /// Typed multipart extractor module,
 /// available with `typed-multipart` feature.
-#[cfg(feature = "typed-multipart")]
+#[cfg(feature = "typed_multipart")]
 pub mod typed;
+
+#[cfg(feature = "typed_multipart")]
+pub use crate::extract::multipart::typed::TypedMultipart;