+++ package/src/lib.rs
@@ -69,4 +69,6 @@
 
 pub mod extract;
 
+pub mod layers;
+
 pub mod response;